requirements.txt
json
paho-mqtt
pymongo
mysql-connector-python
python-dotenv